Removed logic breaking returning of JSON content
There is no documentation for a send function in node's HTTP, so I assume its supposed to be Express response. 
Anyway, I have never seen a call to `.send` with a response object passed in any documentation. On the other hand it breaks passing objects (intended as json responses) to the `.send` method if they have a field called body.